adicionado pageView

+++ lib/components/charts.dart
@@ -2,33 +2,33 @@ import 'package:flutter/material.dart';
 import 'package:syncfusion_flutter_charts/charts.dart';
 
 class Charts extends StatelessWidget {
-  const Charts(List<SheetsData> chartData, String title) :
-    this._chartData = chartData,
-    this._title = title;
-
+  const Charts(List<SheetsData> chartData, String title)
+      : this._chartData = chartData,
+        this._title = title;
 
   final List<SheetsData> _chartData;
   final String _title;
 
   @override
   Widget build(BuildContext context) {
-    return
-
-          SfCircularChart(
-        title: ChartTitle(text: this._title),
-        legend: Legend(isVisible: true),
-        series: <CircularSeries>[
+    return SfCircularChart(
+      title: ChartTitle(text: this._title),
+      legend: Legend(isVisible: true),
+      series: <CircularSeries>[
         DoughnutSeries<SheetsData, String>(
-          dataSource: _chartData,
-          xValueMapper: (SheetsData data,_) => data.answer,
-          yValueMapper: (SheetsData data,_) => data.amount,
-          dataLabelSettings: DataLabelSettings(isVisible: true,))
-      ],);
+            dataSource: _chartData,
+            xValueMapper: (SheetsData data, _) => data.answer,
+            yValueMapper: (SheetsData data, _) => data.amount,
+            dataLabelSettings: DataLabelSettings(
+              isVisible: true,
+            ))
+      ],
+    );
   }
 }
 
-class SheetsData{
+class SheetsData {
   SheetsData(this.answer, this.amount);
   final String answer;
-  final int amount;
+  final double amount;
 }
